Changes in getUser function and in context
-Create new file to move getUser function
- Not expose setUser from the context
