+++ context/user-context.ts
@@ -3,7 +3,6 @@ import { User } from "../types";
 
 interface IUserContext {
   user?: User;
-  setUser: (user: User) => void;
   setAccessToken: (accessToken?: string) => void;
 }
